adding validation and make use of interceptors

+++ Application/Activity/Delete.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using Application.Core;
+using AutoMapper;
 using MediatR;
 using Persistence;
 
@@ -6,11 +7,11 @@ namespace Application.Activity;
 
 public class Delete
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -19,12 +20,16 @@ public class Delete
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
+                if (activity is null) return null;
                 _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
-                return Unit.Value;
+
+                var result = await _context.SaveChangesAsync() > 0;
+                return result
+                    ? Result<Unit>.Success(Unit.Value)
+                    : Result<Unit>.Failure("Error deleting the activity");
             }
         }
     }
